+++ packages/tools/utils/src/array-[any-boolean].spec.js
@@ -1,11 +1,11 @@
 import {strict as assert} from "assert";
 
-import {makeIsContained, makeIsIncluded} from "./array-[any-boolean]";
+import {makeOccursIn, makeIsIncluded} from "./array-[any-boolean]";
 
 describe("Array -> (Any -> Boolean)", function() {
-  describe("makeIsContained", function() {
+  describe("makeOccursIn", function() {
     it("return a function returning true if the passed array is found in the provided array", function() {
-      const isContained = makeIsContained([
+      const isContained = makeOccursIn([
         [1, 2, 3], [1, 2, 3, 4], [5, 6, 7, 6, 5]
       ]);
 
@@ -13,7 +13,7 @@ describe("Array -> (Any -> Boolean)", function() {
       assert.deepStrictEqual(isContained([1, 2]), false);
     });
     it("return a function returning true if the passed object is found in the provided array", function() {
-      const isContained = makeIsContained([
+      const isContained = makeOccursIn([
         {a: 1}, {a: 2}, {a: {b: {c: 3}}}, {b: 1}
       ]);